frontend Notifications: Fix getting warnings outside of cluster context
The notifications were assuming they were always in a cluster contenxt,
which is not the case for when we are at the home view. So in this case
getting the cluster warnings was failing.

This patch also makes the notifications more logical, as now they warn
from a global view instead of warning just for the current cluster being
seen.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>