frontend: Fix success state when adding kubeconfig
The way we were adding clusters previously (before the stateless
clusters) meant we received the configuration when we added a
cluster. After stateless clusters, that same call to add a cluster
in the frontend doesn't return a response. This triggered an
exception which meant the UI assumed there was an issue with setting
up the cluster but that was false.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>