github: Replace minikube with kind in CI
This change replaces the use of minikube in the Build Container and
test / build discover and deploy workflow with the use of kind. We
estimate 60% decrease in run time on a cold run for the "Start Cluster"
step of 1 minute (from ~1m50s to ~45s). Even with a cached image for the
minikube run, we estimate a 25% decrease in run time (from ~1m to ~45s).
This should speed up CI workflows for future changes.

Fixes: #1746

Signed-off-by: Evangelos Skopelitis <eskopelitis@microsoft.com>