backend: Add route and logic for renaming cluster
This adds route that handles logic to rename clusters. When user updates
their cluster name from UI, it sends the customName and source of the
cluster where it is fetched from. The backend checks the source and
updates the context in the source. It also maps the new customName to
original name of cluster in the source.

Fixes: #1653
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>