frontend: Use new Table component in the ResourceTable
Replaces SimpleTable with the new Table component. Out of the box
ResourceTable will enable sorting and filtering on all columns.
To provide that functionality 'getter' in the column definition has
to provide a plaintext value (but will keep working if it returns jsx
for compatibility reasons). e2e and snapshot tests were also updated

Signed-off-by: Oleksandr Dubenko <oldubenko@microsoft.com>