Makefile: Fix run-backend/frontend on win, make faster on linux
This works with Windows, because before setting environment
variables did not work when using cmd or powershell.

Also use nice -19 to speed things up on WSL a lot. Because in a VM
it can be twice as slow normally. It's only getting priority when
compiling, so that is usually what we want.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>