frontend: ReleaseNotes/UpdatePopup: Fix for when internet fails
If there is an error connecting to the github releases, then it
used to be that a message would stay on the screen about fetching
release notes... but it would just stay there.

Now it disapears after it times out, and then a message is shown
that the fetch failed. This message also disapears after 6 seconds.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>