Makefile: github/workflows/frontend: Add i18n-check to CI
Fixes tests so they are not run twice #2125

Running it inside the tests makes the translation files be written
when the tests run, causing the tests to run twice.

Running it as a pre-commit hook is a bit slow (3.0s), because
it looks at all the files that have changed, and then writes all
the translation files again.

So instead we run it in CI so it catches the issues.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>