frontend: Prevent success pop-up on invalid operations
This change ensures that invalid resource operations (e.g. changing a
resource kind) do not trigger a success notification. Previously, the
app would indicate a successful attempt to apply changes even when it is
unsuccessful. Now when trying to perform an invalid apply/edit
operation, instead the editor will reappear with the error message and a
failure notification.

Fixes: #2123

Signed-off-by: Evangelos Skopelitis <eskopelitis@microsoft.com>