app: Use shell to run headlamp-server
This takes care to try extra hard to kill headlamp-server.

Additionally this handles zsh better which does not pass
on environment variables by default.

Because node does not use the users shell by default and instead
uses the /bin/sh this also takes care to use the configured shell.
Meaning that the env variables will be like what is used in the
terminal.

It makes sure that zsh does not hang if it is not setup without
configuration files in interactive mode.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>