frontend KubeConfigLoader: Do not override Next button's color
The button was supposed to be black but the text was also being set
as black. Since primary buttons are now blue from the theme, we
shouldn't hardcode their color like that when it comes to normal
buttons. This patch removes that custom styling from that "next"
button, making it use the default primary color.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>