eslint-config: Add @headlamp-k8s/eslint-config 0.6.0
Make app, frontend, and headlamp-plugin use it, and update
their related packages.

Based on @kinvolk/eslint-config which has been archived.

Also add @mui/icons-material so that packages would resolve now
that mui 6 is out. It had a bad version marker which
tried to pull in mui 6.

Adds table to headlamp-plugin, because it is used to display
table data inside headlamp-plugin but was never specified as
a dependency. However whatever package was using it stopped
so table was no longer being installed.

Had to pin msw to one version, because latest is broken
with a graphql import.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>