frontend: Add check for adding same name clusters
Earlier user were able to submit clusters with same name to the backend,
due to which it was getting overwritten on the proxy. Even if the
cluster was shown in frontend the backend did not knew about the
duplicate cluster.

kubectl throws the error of adding duplicate context name. We are
adding feature for the same.

If now user try to add a context which has the same name as earlier
context the app with throw an error and ask them to change the context
name.

Fixes: #2236
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>