backend: Add websocket multiplexer
This adds websocket multiplexer to the backend. Frontend now make a
single websocket call to the backend. Once that connection is
established with the backend it will send message to the backend with
appropriate data. Backend will open multiple websockets and act as a
proxy to the frontend. It will make request to k8s server and return
data to frontend.

This also adds retry logic if the connection is broken between frontend
and backend.

Fixes: #1802

Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>