backend: Prevent sending refresh signal before the app is ready
The backend was sending a refresh signal to the frontend very early due
to reading the plugins folders, and this resulted in a refresh right
after the UI was loaded (desktop only).

It is only interesting for the app to refresh after a plugin change if
it already has the list of plugins, so these changes prevent that refresh
until the app requests the plugins list.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>