frontend AuthChooser: Set testing auth as true by default
We should only set the testing state as false when we go through the
testing. This prevents the case of the test happening so quickly that
the UI is rendered as never testing the auth, which means the options of
using a token are shown to the user when they are being redirected to
the actual cluster after a successful auth.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>