frontend: Add websocket multiplexer
This adds a single websocket connection from frontend to backend and
sends lists of messages to the backend. With the help of messages
backend creates multiple websocket connection to k8s API and returns
back the data.

This solves the issue for limiting of websocket connection in frontend
in case of multi cluster setup.

Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>