headlamp-build: Add package.json for scripts from Makefile
This is to simplify using Headlamp so that people do not need to
have installed gnu Make. Which isn't on Windows, and isn't on
some Mac/unix systems.

Just starting with the start script, the rest can come later.

This package is not intended to be published, and so the name is
not important at this stage.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>