backend: Prevent flashing terminal on Windows
On Windows, even when the exec plugin's command is not interactive,
running the command was still making the terminal window flash (default
behavior on Windows). This can be annoying if a tool runs the exec
plugin frequently. This patch prevents that problem by setting a sys proc
attribute to hide windows, only when in the Windows platform and when the
exec is non-interactive.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>