backend: Do not include clusters with errors when loading them
We were loading clusters and checking if they had errors, but we'd
still include them in the final list.

This patch fundamentally changes how loadContextsFromData works, since
it seemed like it should be returning all contexts regardless of
whether they had errors or not, but some other parts of the code were
counting on these contexts having no errors. After this commit, only
contexts without errors are returned in that list.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>